pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image-translate-backend"
version = "0.1.0"
description = "FastAPI backend service for image translation using Gemini API"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = "==3.11.*"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "gunicorn==21.2.0",
    "redis==5.2.0",
    "aiofiles==23.2.1",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "pydantic-settings==2.1.0",
    "loguru==0.7.2",
    "google-genai==0.3.0",
    "Pillow==10.1.0",
    "pydantic==2.5.0",
    "httpx==0.25.2",
    "python-magic==0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "pytest-mock"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio", 
    "httpx",
    "pytest-mock"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"